[Fix] #472 - tagType의 스타일 관련 enum들 model로 이전
