+++ SOPT-iOS/Projects/Data/Sources/Transform/HomeGroupPostsTransform.swift
@@ -13,6 +13,12 @@ import Networks
 
 extension HomeGroupEntity {
     public func toDomain() -> HomeGroupPostModel {
-        return HomeGroupPostModel(id: id, title: title, category: category, canJoinOnlyActiveGeneration: canJoinOnlyActiveGeneration, joinableParts: joinableParts, canJoinAllParts: canJoinAllParts, status: status, imageUrl: imageUrl)
+        return HomeGroupPostModel(id: id,
+                                  title: title,
+                                  category: HomeGroupPostModel.Category(rawValue: category) ?? .event,
+                                  canJoinOnlyActiveGeneration: canJoinOnlyActiveGeneration,
+                                  joinableParts: joinableParts,
+                                  canJoinAllParts: canJoinAllParts,
+                                  status: HomeGroupPostModel.Status(rawValue: status) ?? .applyAble, imageUrl: imageUrl)
     }
 }
