+++ SOPT-iOS/Projects/Domain/Sources/Model/HomeGroupPostModel.swift
@@ -6,21 +6,22 @@
 //  Copyright © 2025 SOPT-iOS. All rights reserved.
 //
 
-import Foundation
+import UIKit
 
-import Networks
+import Core
+import DSKit
 
 public struct HomeGroupPostModel: Codable {
     public let id: Int
     public let title: String
-    public let category: String
+    public let category: Category
     public let canJoinOnlyActiveGeneration: Bool
     public let joinableParts: [String]
     public let canJoinAllParts: Bool
-    public let status: HomeGroupEntity.Status
+    public let status: Status
     public let imageUrl: String
     
-    public init(id: Int, title: String, category: String, canJoinOnlyActiveGeneration: Bool, joinableParts: [String], canJoinAllParts: Bool, status: HomeGroupEntity.Status, imageUrl: String) {
+    public init(id: Int, title: String, category: Category, canJoinOnlyActiveGeneration: Bool, joinableParts: [String], canJoinAllParts: Bool, status: Status, imageUrl: String) {
         self.id = id
         self.title = title
         self.category = category
@@ -29,5 +30,71 @@ public struct HomeGroupPostModel: Codable {
         self.canJoinAllParts = canJoinAllParts
         self.status = status
         self.imageUrl = imageUrl
+    }
+    
+    // MARK: - Category Enum
+    
+    public enum Category: String, Codable {
+        case event = "EVENT"
+        case study = "STUDY"
+        
+        public var text: String {
+            switch self {
+            case .event:
+                return "행사"
+            case .study:
+                return "스터디"
+            }
+        }
+        
+        public var textColor: UIColor {
+            switch self {
+            case .event:
+                return DSKitAsset.Colors.success.color
+            case .study:
+                return DSKitAsset.Colors.secondary.color
+            }
+        }
+    }
+    
+    // MARK: - Status Enum
+    
+    public enum Status: String, Codable {
+        case beforeStart = "BEFORE_START"
+        case applyAble = "APPLY_ABLE"
+        case recruitmentComplete = "RECRUITMENT_COMPLETE"
+        
+        public var text: String {
+            switch self {
+            case .beforeStart:
+                return I18N.Home.Group.beforeStart
+            case .applyAble:
+                return I18N.Home.Group.applyAble
+            case .recruitmentComplete:
+                return I18N.Home.Group.recruitmentComplete
+            }
+        }
+        
+        public var textColor: UIColor {
+            switch self {
+            case .beforeStart:
+                return DSKitAsset.Colors.gray800.color
+            case .applyAble:
+                return DSKitAsset.Colors.gray800.color
+            case .recruitmentComplete:
+                return DSKitAsset.Colors.gray800.color
+            }
+        }
+        
+        public var backgroundColor: UIColor {
+            switch self {
+            case .beforeStart:
+                return DSKitAsset.Colors.attention.color
+            case .applyAble:
+                return DSKitAsset.Colors.secondary.color
+            case .recruitmentComplete:
+                return DSKitAsset.Colors.gray100.color
+            }
+        }
     }
 }
